Moved the Position and PlayerType to Player folder. Created a servers property for the User entity to know what server(s) the player is on
