+++ src/main/java/com/league_buddies/backend/user/PlayerType.java
@@ -1,4 +1,4 @@
-package com.league_buddies.backend.playerType;
+package com.league_buddies.backend.user;
 
 public enum PlayerType {
     Competitive,
