+++ src/main/java/com/league_buddies/backend/user/Position.java
@@ -1,4 +1,4 @@
-package com.league_buddies.backend.position;
+package com.league_buddies.backend.user;
 
 public enum Position {
     TOP,
