+++ src/main/java/com/league_buddies/backend/user/Server.java
@@ -1,0 +1,10 @@
+package com.league_buddies.backend.user;
+
+public enum Server {
+    NA,
+    EUNE,
+    EUW,
+    LAN,
+    OCE,
+    LAS
+}
