+++ src/main/java/com/league_buddies/backend/user/UserService.java
@@ -60,6 +60,7 @@ public class UserService {
             currUser.setPlayerType(user.getPlayerType());
             currUser.setWinRate(user.getWinRate());
             currUser.setRole(user.getRole());
+            currUser.setServers(user.getServers());
             userRepository.save(currUser);
             return currUser;
         }
